Fix campaign search using search_campaigns function
