+++ package.json
@@ -46,7 +46,7 @@
     "fontkit": "^2.0.4",
     "input-otp": "^1.4.2",
     "lucide-react": "^0.542.0",
-    "next": "^16.0.2-canary.9",
+    "next": "15.5.2",
     "next-themes": "^0.4.6",
     "pdfkit": "^0.17.2",
     "react": "19.1.0",
