+++ src/lib/supabase/server.ts
@@ -109,5 +109,3 @@ export const getOrganizationProfile = cache(
     return { email: user.email, ...profile };
   }
 );
-export { createClient };
-
